TSify utils/StyledLink
